+++ pages/order-history.tsx
@@ -1,0 +1,38 @@
+import { useEffect, useState } from 'react'
+
+import Head from 'next/head'
+import Header from '../components/Header'
+import { Order, getMyOrders } from '../services/orders'
+
+export default function OrderHistory() {
+  const [loading, setLoading] = useState<boolean>(false)
+  const [orders, setOrders] = useState<Array<Order>>([])
+
+  function handleGetOrders() {
+    setLoading(true)
+    getMyOrders()
+      .then((data) => setOrders(data.docs))
+      .catch((err) => console.log(err))
+      .finally(() => setLoading(false))
+  }
+
+  useEffect(() => {
+    handleGetOrders()
+  }, [])
+
+  return (
+    <>
+      <Head>
+        <title>Eurocambiovla - Historial</title>
+        <meta name="description" content="Generated by create next app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <div className="container">
+        <Header />
+        {loading ? <div>Loading...</div> : orders.map((o: Order) => <div key={o.id}>
+            <p>id: {o.id}</p>
+        </div>)}
+      </div>
+    </>
+  )
+}
