+++ utils/PayMethods.ts
@@ -1,0 +1,5 @@
+export enum PayMethods {
+    PAYPAL = 'PayPal',
+    CARD = 'Card',
+    TRANSFER = 'Transferencia'
+}
