+++ utils/bankOptions.ts
@@ -1,0 +1,102 @@
+export const bankOptions = [
+  {
+    label: '100% Banco',
+    value: '100% Banco'
+  },
+  {
+    label: 'Banco Activo',
+    value: 'Banco Activo'
+  },
+  {
+    label: 'Banco Bicentenario',
+    value: 'Banco Bicentenario'
+  },
+  {
+    label: 'Banco Caroní',
+    value: 'Banco Caroní'
+  },
+  {
+    label: 'Banco de Venezuela',
+    value: 'Banco de Venezuela'
+  },
+  {
+    label: 'Banco del Sur',
+    value: 'Banco del Sur'
+  },
+  {
+    label: 'Banco del Tesoro',
+    value: 'Banco del Tesoro'
+  },
+  {
+    label: 'Banco Exterior',
+    value: 'Banco Exterior'
+  },
+  {
+    label: 'Banco Mercantil',
+    value: 'Banco Mercantil'
+  },
+  {
+    label: 'Bancamiga',
+    value: 'Bancamiga'
+  },
+  {
+    label: 'Bancaribe',
+    value: 'Bancaribe'
+  },
+  {
+    label: 'Bancrecer',
+    value: 'Bancrecer'
+  },
+  {
+    label: 'Bangente',
+    value: 'Bangente'
+  },
+  {
+    label: 'Banesco',
+    value: 'Banesco'
+  },
+  {
+    label: 'BFC - Banco Fondo Común',
+    value: 'BFC - Banco Fondo Común'
+  },
+  {
+    label: 'BOD - Banco Occidental de Descuento',
+    value: 'BOD - Banco Occidental de Descuento'
+  },
+  {
+    label: 'Bancos Venezolano de Crédito',
+    value: 'Bancos Venezolano de Crédito'
+  },
+  {
+    label: 'Banplus',
+    value: 'Banplus'
+  },
+  {
+    label: 'Banplus',
+    value: 'Banplus'
+  },
+  {
+    label: 'Banco Plaza',
+    value: 'Banco Plaza'
+  },
+  {
+    label: 'Banco Sofitasa',
+    value: 'Banco Sofitasa'
+  },
+  {
+    label: 'BBVA Provincial',
+    value: 'BBVA Provincial'
+  },
+  {
+    label: 'BNC - Banco Nacional de Crédito',
+    value: 'BNC - Banco Nacional de Crédito'
+  },
+  {
+    label: 'Citibank Sucursal Venezuela',
+    value: 'Citibank Sucursal Venezuela'
+  },
+  {
+    label: 'Mi Banco',
+    value: 'Mi Banco'
+  }
+]
